+++ src/components/Calculators/Electronics/Pcb/TrackCurrentIpc2221A/MainView.vue
@@ -104,7 +104,7 @@
 <script>
 
   import Calc from 'src/misc/CalculatorEngineV2/Calc'
-  import {CalcVarNumeral} from 'src/misc/CalculatorEngineV2/CalcVarNumeral'
+  import {CalcVarNumeric} from 'src/misc/CalculatorEngineV2/CalcVarNumeric'
   import {CalcVarComboBox} from 'src/misc/CalculatorEngineV2/CalcVarComboBox'
   import {UnitMulti} from 'src/misc/CalculatorEngineV2/UnitMulti'
   import {UnitFunc} from 'src/misc/CalculatorEngineV2/UnitFunc'
@@ -130,7 +130,7 @@
       // ============================================ //
       // ============ TRACK CURRENT (input) ========= //
       // ============================================ //
-      calc.addVar(new CalcVarNumeral({
+      calc.addVar(new CalcVarNumeric({
         name: 'trackCurrent',
         typeEqn: () => {
           return 'input'
@@ -164,7 +164,7 @@
       // ============================================ //
       // ============= TEMP. RISE (input) =========== //
       // ============================================ //
-      calc.addVar(new CalcVarNumeral({
+      calc.addVar(new CalcVarNumeric({
         name: 'tempRise',
         typeEqn: () => {
           return 'input'
@@ -214,7 +214,7 @@
       // ============================================ //
       // ========== TRACK THICKNESS (input) ========= //
       // ============================================ //
-      calc.addVar(new CalcVarNumeral({
+      calc.addVar(new CalcVarNumeric({
         name: 'trackThickness',
         typeEqn: () => {
           return 'input'
@@ -272,7 +272,7 @@
       // ============================================================================================= //
       // ===================================== MIN. TRACK WIDTH (output) ============================= //
       // ============================================================================================= //
-      calc.addVar(new CalcVarNumeral({
+      calc.addVar(new CalcVarNumeric({
         name: 'minTrackWidth',
         typeEqn: () => {
           return 'output'